+++ src/Application/Command/Planner/PrePlanningProcess/TransformPriorityRequestsIntoMeetingHandler.php
@@ -1,0 +1,174 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Command\Planner\PrePlanningProcess;
+
+use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
+use Proximum\Vimeet\Application\Command\Meeting\TransformRequestIntoMeeting;
+use Proximum\Vimeet\Domain\Model\Meeting;
+use Proximum\Vimeet\Domain\Model\Rule;
+use Proximum\Vimeet\Domain\Model\Type;
+use Proximum\Vimeet\Domain\Planner\ExportSolutionType;
+use Proximum\Vimeet\Domain\Repository\Meeting\RequestRepositoryInterface;
+use Proximum\Vimeet\Domain\Repository\RuleRepositoryInterface;
+
+class TransformPriorityRequestsIntoMeetingHandler
+{
+    /** @var RuleRepositoryInterface */
+    private $ruleRepository;
+
+    /** @var RequestRepositoryInterface */
+    private $requestRepository;
+
+    /** @var CommandBusInterface */
+    private $commandBus;
+
+    public function __construct(
+        CommandBusInterface $commandBus,
+        RequestRepositoryInterface $requestRepository,
+        RuleRepositoryInterface $ruleRepository
+    ) {
+        $this->commandBus = $commandBus;
+        $this->requestRepository = $requestRepository;
+        $this->ruleRepository = $ruleRepository;
+    }
+
+    /**
+     * @param TransformPriorityRequestsIntoMeeting $command
+     *
+     * @return Meeting[]
+     */
+    public function handle(TransformPriorityRequestsIntoMeeting $command): array
+    {
+        if (ExportSolutionType::SOLUTION_FROM_SCRATCH === $command->solutionType) {
+            return [];
+        }
+
+        $event = $command->event;
+
+        $approvedRequests = $this->requestRepository->findApprovedAndPrioritizedWithoutMeeting($event);
+
+        $orderedRequests = $this->getOrderedRequestsByPriorities($event, $approvedRequests);
+
+        $createdMeetings = [];
+        foreach ($orderedRequests as $request) {
+            $createdMeetings[] = $this->commandBus->handle(
+                new TransformRequestIntoMeeting($request, Meeting::CREATED_BY_PLANNER, true, true)
+            );
+        }
+
+        return $createdMeetings;
+    }
+
+    /**
+     * @param \Proximum\Vimeet\Domain\Model\Event $event
+     * @param Meeting\Request[]                   $approvedRequests
+     *
+     * @return Meeting\Request[]
+     */
+    private function getOrderedRequestsByPriorities(
+        \Proximum\Vimeet\Domain\Model\Event $event,
+        $approvedRequests
+    ): array {
+        $wswRules = $this->ruleRepository->getByEvent($event);
+
+        usort(
+            $wswRules,
+            static function (Rule $ruleA, Rule $ruleB) {
+                return $ruleA->getPriority() <=> $ruleB->getPriority();
+            }
+        );
+
+        $orderedRequests = [];
+
+        foreach ($wswRules as $rule) {
+            // double priority
+            $toTreatRequest = [];
+            foreach ($approvedRequests as $i => $approvedRequest) {
+                if ($this->isRequestConcernedByRule($rule, $approvedRequest)
+                    && $approvedRequest->isFromPriority() && $approvedRequest->isToPriority()) {
+                    $assignedSheetId = $approvedRequest->getFromSheet()->getId();
+
+                    $toTreatRequest[$assignedSheetId][] = $approvedRequest;
+                    unset($approvedRequests[$i]);
+                }
+            }
+
+            $orderedRequests = $this->sortBySheetAndPushRequests($toTreatRequest, $orderedRequests);
+
+            // single priority
+            $toTreatRequest = [];
+            foreach ($approvedRequests as $i => $approvedRequest) {
+                if ($this->isRequestConcernedByRule($rule, $approvedRequest)) {
+                    $assignedSheetId = $approvedRequest->isFromPriority()
+                        ? $approvedRequest->getFromSheet()->getId() : $approvedRequest->getToSheet()->getId();
+
+                    $toTreatRequest[$assignedSheetId][] = $approvedRequest;
+                    unset($approvedRequests[$i]);
+                }
+            }
+
+            $orderedRequests = $this->sortBySheetAndPushRequests($toTreatRequest, $orderedRequests);
+        }
+
+        // not treated requests - double priority
+        $toTreatRequest = [];
+        foreach ($approvedRequests as $i => $approvedRequest) {
+            if ($approvedRequest->isFromPriority() && $approvedRequest->isToPriority()) {
+                $assignedSheetId = $approvedRequest->getFromSheet()->getId();
+
+                $toTreatRequest[$assignedSheetId][] = $approvedRequest;
+                unset($approvedRequests[$i]);
+            }
+        }
+
+        $orderedRequests = $this->sortBySheetAndPushRequests($toTreatRequest, $orderedRequests);
+
+        // not treated requests - single priority
+        $toTreatRequest = [];
+        foreach ($approvedRequests as $i => $approvedRequest) {
+            $assignedSheetId = $approvedRequest->isFromPriority()
+                ? $approvedRequest->getFromSheet()->getId() : $approvedRequest->getToSheet()->getId();
+
+            $toTreatRequest[$assignedSheetId][] = $approvedRequest;
+            unset($approvedRequests[$i]);
+        }
+
+        $orderedRequests = $this->sortBySheetAndPushRequests($toTreatRequest, $orderedRequests);
+
+        return $orderedRequests;
+    }
+
+    private function isRequestConcernedByRule(Rule $wswRule, Meeting\Request $approvedRequest): bool
+    {
+        $typeSeer = $approvedRequest->getFromSheet()->getType();
+        $typeSeable = $approvedRequest->getToSheet()->getType();
+
+        return ($wswRule->getSeer() instanceof Type
+                ? $wswRule->getSeer() === $typeSeer
+                : $wswRule->getSeer()->getTypes()->contains($typeSeer)) &&
+            ($wswRule->getSeeable() instanceof Type
+                ? $wswRule->getSeeable() === $typeSeable
+                : $wswRule->getSeeable()->getTypes()->contains($typeSeable));
+    }
+
+    /**
+     * @param array             $toTreatRequest
+     * @param Meeting\Request[] $orderedRequests
+     *
+     * @return Meeting\Request[]
+     */
+    private function sortBySheetAndPushRequests(array $toTreatRequest, array $orderedRequests): array
+    {
+        while (\count($toTreatRequest)) {
+            foreach ($toTreatRequest as $sheetId => $requests) {
+                $orderedRequests[] = array_shift($requests);
+                $toTreatRequest[$sheetId] = $requests;
+                if (empty($requests)) {
+                    unset($toTreatRequest[$sheetId]);
+                }
+            }
+        }
+
+        return $orderedRequests;
+    }
+}