+++ tests/Application/Command/Planner/PrePlanningProcess/PrePlanningProcessHandlerTest.php
@@ -7,16 +7,29 @@ use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
 use Proximum\Vimeet\Application\Command\Planner\PrePlanningProcess\TransformApprovedRequestsByLinkedSheetsIntoMeeting;
 use Proximum\Vimeet\Application\Command\Planner\PrePlanningProcess\PrePlanningProcess;
 use Proximum\Vimeet\Application\Command\Planner\PrePlanningProcess\PrePlanningProcessHandler;
-use Proximum\Vimeet\Application\Command\Planner\PrePlanningProcess\RequestAcceptedByAllLinkedSheets;
+use Proximum\Vimeet\Application\Command\Planner\PrePlanningProcess\TransformPriorityRequestsIntoMeeting;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Meeting;
 
 class PrePlanningProcessHandlerTest extends TestCase
 {
-    public function testHandle()
+    /**
+     * TransformPriorityRequestsIntoMeetingHandler should be called before
+     *      TransformApprovedRequestsByLinkedSheetsIntoMeetingHandler.
+     * But it's against Prophecy philosophy to test executions order.
+     *
+     * See more at https://github.com/phpspec/prophecy/issues/130
+     */
+    public function testHandle(): void
     {
         $event = $this->prophesize(Event::class);
         $commandBus = $this->prophesize(CommandBusInterface::class);
+
+        $commandBus
+            ->handle(new TransformPriorityRequestsIntoMeeting($event->reveal(), Meeting::CREATED_BY_PLANNER))
+            ->shouldBeCalled()
+        ;
+
         $commandBus
             ->handle(new TransformApprovedRequestsByLinkedSheetsIntoMeeting($event->reveal(), Meeting::CREATED_BY_PLANNER))
             ->shouldBeCalled()